move handleGlobalEnterQuickSubmit into a separate file to avoid cycle-import (#20679)
